uint16 for no overflow

+++ DigitDetection.py
@@ -70,4 +70,4 @@ class DigitDetection(Thread):
     def __init__(self):
         super(DigitDetection, self).__init__()
         self.cancelled = False
-        self.number = np.array([0, 0, 0, 0, 0], np.uint8)
+        self.number = np.array([0, 0, 0, 0, 0], np.uint16)
